save func to user model added create user and register combined urls and middleware fixed
