+++ lead/urls.py
@@ -1,12 +1,10 @@
 from django.urls import path
-from .views import payment_list, create_payment, update_payment, balance_report, create_lead_view, \
-    create_user_view, change_lead_admin_view, change_student_admin_view, lead_list_view, \
+from .views import payment_list, create_payment, update_payment, balance_report, create_lead_view, change_lead_admin_view, change_student_admin_view, lead_list_view, \
     student_list_view, admin_lead_view, lead_update_view, create_student_view, my_students_list_view, \
     student_update_view, student_detail, create_student
 
 urlpatterns = [
     path('create_lead/', create_lead_view, name='create-lead'),
-    path('create_user/', create_user_view, name='create-user'),
     path('create_student/', create_student_view, name='create-student'),
     path('change_lead_admin/<int:lead_id>/', change_lead_admin_view, name='change-lead-admin'),
     path('change_student_admin/<int:student_id>/', change_student_admin_view, name='change-student-admin'),
